continuando o front e preparando alguns ajustes
